src/data_collection.py
"""
Data Collection Module for Emotion Recognition System

This module handles data collection from external sources and APIs
for the emotion recognition system.
"""

import requests
import pandas as pd
import numpy as np
from PIL import Image
import os
import json
from typing import List, Dict, Optional
import time

class EmotionDataCollector:
    """
    Collects emotion data from various external sources
    """
    
    def __init__(self, api_key: Optional[str] = None):
        self.api_key = api_key
        self.base_url = "https://api.example.com"  # Replace with actual API
        
    def collect_from_api(self, endpoint: str, params: Dict) -> Dict:
        """
        Collect data from external API endpoint
        
        Args:
            endpoint: API endpoint URL
            params: Query parameters
            
        Returns:
            API response data
        """
        try:
            headers = {"Authorization": f"Bearer {self.api_key}"} if self.api_key else {}
            response = requests.get(f"{self.base_url}/{endpoint}", 
                                  params=params, headers=headers, timeout=30)
            response.raise_for_status()
            return response.json()
        except requests.RequestException as e:
            print(f"Error collecting data from API: {e}")
            return {}
    
    def collect_emotion_samples(self, num_samples: int = 100) -> List[Dict]:
        """
        Collect emotion samples from external source
        
        Args:
            num_samples: Number of samples to collect
            
        Returns:
            List of emotion sample data
        """
        samples = []
        
        # Simulate data collection from external source
        emotions = ['happy', 'sad', 'angry', 'surprise', 'fear', 'disgust', 'neutral']
        
        for i in range(num_samples):
            sample = {
                'id': f"sample_{i:04d}",
                'emotion': np.random.choice(emotions),
                'confidence': np.random.uniform(0.7, 0.95),
                'timestamp': time.time(),
                'source': 'external_api'
            }
            samples.append(sample)
        
        return samples
    
    def save_collected_data(self, data: List[Dict], filename: str):
        """
        Save collected data to file
        
        Args:
            data: Collected data
            filename: Output filename
        """
        df = pd.DataFrame(data)
        df.to_csv(filename, index=False)
        print(f"Data saved to {filename}")
    
    def validate_data_quality(self, data: List[Dict]) -> Dict:
        """
        Validate quality of collected data
        
        Args:
            data: Collected data
            
        Returns:
            Quality metrics
        """
        if not data:
            return {"quality_score": 0, "issues": ["No data collected"]}
        
        df = pd.DataFrame(data)
        
        quality_metrics = {
            "total_samples": len(df),
            "unique_emotions": df['emotion'].nunique(),
            "avg_confidence": df['confidence'].mean(),
            "min_confidence": df['confidence'].min(),
            "quality_score": 0.95  # Simulated quality score
        }
        
        return quality_metrics

def main():
    """
    Main function for data collection
    """
    print("Starting emotion data collection...")
    
    # Initialize collector
    collector = EmotionDataCollector()
    
    # Collect samples
    samples = collector.collect_emotion_samples(num_samples=50)
    
    # Validate data quality
    quality = collector.validate_data_quality(samples)
    print(f"Data quality metrics: {quality}")
    
    # Save data
    collector.save_collected_data(samples, "data/collected_emotions.csv")
    
    print("Data collection completed!")

if __name__ == "__main__":
    main()
