requirements.txt
streamlit==1.49.1
deepface==0.0.95
pillow==11.2.1
numpy==2.1.3
tensorflow==2.20.0
tf-keras==2.20.1
opencv-python==4.12.0.88
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.2
jupyter==1.1.1
requests==2.32.3
